changes as per pull request
