+++ package.json
@@ -19,11 +19,11 @@
   "author": "joshramsbottom",
   "license": "MIT",
   "dependencies": {
-    "babel-cli": "^6.24.0",
-    "babel-preset-env": "^1.3.2",
-    "discord.js": "^11.0.0",
-    "dotenv": "^5.0.0",
-    "ghastly": "^0.4.0"
+    "babel-cli": "^6.26.0",
+    "babel-preset-env": "^1.6.1",
+    "discord.js": "^11.3.2",
+    "dotenv": "^5.0.1",
+    "ghastly": "^0.4.1"
   },
   "devDependencies": {
     "nodemon": "^1.14.8"
