+++ src/services/PugQueue.js
@@ -1,6 +1,4 @@
-import { getFullName } from '../util'
-import { getRoles } from '../util'
-import { pug_map } from '../util'
+import { getFullName, getRoleEmoji, getRandomMap } from '../util'
 
 export class PugQueue {
   constructor () {
@@ -15,7 +13,7 @@ export class PugQueue {
 
   add (member) {
     const name = getFullName(member)
-    const pug_roles = getRoles(member)
+    const roleEmoji = getRoleEmoji(member)
 
     if (this.queue.includes(member)) {
       return `${name} is already queued.`
@@ -28,37 +26,36 @@ export class PugQueue {
     this.queue.push(member)
     member.addRole(process.env.PUGS_ROLE)
 
-    return `${String.fromCodePoint(0x2705)} ${name} ${pug_roles} added to queue. ${this.getQueueState()}`
+    return `${String.fromCodePoint(0x2705)} ${name} ${roleEmoji} added to queue. ${this.getQueueState()}`
   }
 
   attemptGameStart (guild) {
-    if (this.queue.length < process.env.TEAM_SIZE * 2) {
+    if (this.queue.length < process.env.TEAM_SIZE * 1) {
       return
     }
 
     let mentions = ``
     this.queue.forEach(member => {
-      const name = getFullName(member)
-      const pug_roles = getRoles(member)
-      mentions += `${member} ${pug_roles}\n`
+      const roleEmoji = getRoleEmoji(member)
+      mentions += `${member} ${roleEmoji}\n`
     })
 
     // Announce game start and alert players
-    const announce_channel = guild.client.channels.get(process.env.PUGS_ANNOUNCEMENTCHANNEL)
-    const map = pug_map()
-    announce_channel.send(`❮❮❮\t\t **Match Starting!**\t\t ❯❯❯\n
+    const announceChannel = guild.client.channels.get(process.env.PUGS_ANNOUNCEMENTCHANNEL)
+    const map = getRandomMap()
+    announceChannel.send(`❮❮❮\t\t **Match Starting!**\t\t ❯❯❯\n
 *Following players head over to Match Draft:*\n
 ${mentions}\n
 
 ${map}
 
 ❯❯❯\t\t*end of match announcement*\t\t❮❮❮\n
-	`)
-    
+`)
+
     // Notification for queue full
-    const pugs_channel = guild.client.channels.get(process.env.PUGS_CHANNEL)
-    pugs_channel.send(`We have our 12 players, lets go!`)
-    
+    const pugsChannel = guild.client.channels.get(process.env.PUGS_CHANNEL)
+    pugsChannel.send(`We have our 12 players, lets go!`)
+
     // Empty queue
     this.queue.length = 0
   }
