+++ src/util/index.js
@@ -1,3 +1,2 @@
 export * from './utils'
-export * from './pug_roles'
-export * from './pug_map'
+export * from './getRandomMap'
